+++ changelog.md
@@ -8,6 +8,7 @@ CHANGE LOG
 
 - Default nginx configuration now redirect to backend when admin or api url does not have trailing /.
 - Category ordering in admin panel
+- Add `import_data` command to import data from old LizMap GeoJSON file.
 
 
 1.0.0        (2024-07-19)